+++ Light Vox Engine/stdafx.h
@@ -4,18 +4,18 @@
 //For windows and graphics
 #include <Windows.h>
 #include <wrl.h>
-#include "d3dx12.h"
+#include "Utils/d3dx12.h"
 #include <dxgi1_4.h>
 #include <d3dcompiler.h>
 #include <DirectXMath.h>
-
+#include <stdlib.h>  
 //events
 #include <pix3.h>
 
 //utils
-#include "lvUtils.h"
-#include "DXSampleHelper.h"
-#include "Configs.h"
+#include "Utils/lvUtils.h"
+#include "Utils/DXSampleHelper.h"
+#include "Utils/Configs.h"
 #include <cstdint>
 
 //Math 
@@ -29,7 +29,7 @@
 //Debug
 #if defined(_DEBUG)
 #define _CRTDBG_MAP_ALLOC  
-#include <stdlib.h>  
+
 #include <crtdbg.h> 
 
 #define DEBUG_NEW new(_NORMAL_BLOCK, __FILE__, __LINE__)
